[OpenMP][NFC] Separate OpenMP/OpenACC specific mapping code
While this does not really encapsulate the mapping code, it at least
moves most of the declarations out of the way.
